Added support for the --retry-interval command line option. It is intended as a broker exception retry interval and is specified as number of seconds.
By default, if there is an exception while communicating
with the broker, celery-exporter exits with a stack trace.

It is not uncommon to temporarily lose connectivity to redis
for a very short amount of time. With the default behavior, the
exporter will need a restart thus losing all accumulated
metrics.

With the new --retry-interval options specified, one can
specify and optional retry interval (i.e. 5 seconds) and
celery-exporter will restart the celery EventReceiver
for that broker.

The default is 0, (i.e. do not retry) in order to keep
compatibility (and expected behavior) with the current
celery-exporter.